fixing bugs #2 and #3
Updated BufferedSerialInputStream to optionally throw an IOException
when an InterruptedException is thrown.  Added a function on the
SerialPort to set this functionality.  Fixed the inputstream returning 0
when the serial port is closed(at least on Windows), will now return -1.
Needs to be tested on Linux.  POM should no longer try to sign by
default, only when we release.